+++ Cargo.toml
@@ -11,6 +11,7 @@ members = [
     "crates/core",
     "crates/model-executor",
     "crates/runtime",
+    "crates/agent-runtime",
     "crates/mcp",
     "crates/hooks",
     "crates/permission",
@@ -61,6 +62,7 @@ zene-tools = { path = "crates/tools" }
 zene-mcp = { path = "crates/mcp" }
 zene-model-executor = { path = "crates/model-executor" }
 zene-runtime = { path = "crates/runtime" }
+zene-agent-runtime = { path = "crates/agent-runtime" }
 zene-hooks = { path = "crates/hooks" }
 zene-permission = { version = "0.1.11", path = "crates/permission" }
 zene-tool-runtime = { path = "crates/tool-runtime" }
