+++ crates/agent-runtime/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "zene-agent-runtime"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+publish = false
+
+[dependencies]
+anyhow.workspace = true
+async-trait.workspace = true
+chrono.workspace = true
+parking_lot.workspace = true
+tokio.workspace = true
+tokio-util.workspace = true
+zene-core = { workspace = true }
+zene-permission = { workspace = true }
+zene-runtime = { workspace = true }
+zene-session = { workspace = true }
+zene-turn = { workspace = true }
+
+[dev-dependencies]
+tempfile = "3"
+zene-config = { workspace = true }
+zene-sandbox = { workspace = true }
